+++ src/providers/irc/IrcMessageBuilder.cpp
@@ -74,6 +74,7 @@ void IrcMessageBuilder::appendUsername()
 
     QString username = this->userName;
     this->message().loginName = username;
+    this->message().displayName = username;
 
     // The full string that will be rendered in the chat widget
     QString usernameText = username;
@@ -85,7 +86,7 @@ void IrcMessageBuilder::appendUsername()
 
     this->emplace<TextElement>(usernameText, MessageElementFlag::Username,
                                this->usernameColor_, FontStyle::ChatMediumBold)
-        ->setLink({Link::UserInfo, this->message().displayName});
+        ->setLink({Link::UserInfo, this->message().loginName});
 }
 
 }  // namespace chatterino