+++ src/providers/twitch/api/Helix.hpp
@@ -2,6 +2,7 @@
 
 #include "common/Aliases.hpp"
 #include "common/NetworkRequest.hpp"
+#include "providers/twitch/TwitchEmotes.hpp"
 
 #include <QJsonArray>
 #include <QString>
@@ -267,10 +268,31 @@ struct HelixCheermoteSet {
 struct HelixEmoteSetData {
     QString setId;
     QString ownerId;
+    QString emoteType;
 
     explicit HelixEmoteSetData(QJsonObject jsonObject)
         : setId(jsonObject.value("emote_set_id").toString())
         , ownerId(jsonObject.value("owner_id").toString())
+        , emoteType(jsonObject.value("emote_type").toString())
+    {
+    }
+};
+
+struct HelixChannelEmote {
+    const QString emoteId;
+    const QString name;
+    const QString type;
+    const QString setId;
+    const QString url;
+
+    explicit HelixChannelEmote(QJsonObject jsonObject)
+        : emoteId(jsonObject.value("id").toString())
+        , name(jsonObject.value("name").toString())
+        , type(jsonObject.value("emote_type").toString())
+        , setId(jsonObject.value("emote_set_id").toString())
+        , url(QString(TWITCH_EMOTE_TEMPLATE)
+                  .replace("{id}", this->emoteId)
+                  .replace("{scale}", "3.0"))
     {
     }
 };
@@ -413,6 +435,12 @@ public:
     void getEmoteSetData(QString emoteSetId,
                          ResultCallback<HelixEmoteSetData> successCallback,
                          HelixFailureCallback failureCallback);
+
+    // https://dev.twitch.tv/docs/api/reference#get-channel-emotes
+    void getChannelEmotes(
+        QString broadcasterId,
+        ResultCallback<std::vector<HelixChannelEmote>> successCallback,
+        HelixFailureCallback failureCallback);
 
     void update(QString clientId, QString oauthToken);