+++ src/widgets/helper/ChannelView.cpp
@@ -1928,7 +1928,7 @@ void ChannelView::addContextMenuItems(
         crossPlatformCopy(copyString);
     });
 
-    // Open in new split.
+    // If is a link to a twitch user/stream
     if (hoveredElement->getLink().type == Link::Url)
     {
         static QRegularExpression twitchChannelRegex(
@@ -1947,7 +1947,22 @@ void ChannelView::addContextMenuItems(
         {
             menu->addSeparator();
             menu->addAction("Open in new split", [twitchUsername, this] {
-                this->joinToChannel.invoke(twitchUsername);
+                this->openChannelIn.invoke(twitchUsername,
+                                           FromTwitchLinkOpenChannelIn::Split);
+            });
+            menu->addAction("Open in new tab", [twitchUsername, this] {
+                this->openChannelIn.invoke(twitchUsername,
+                                           FromTwitchLinkOpenChannelIn::Tab);
+            });
+
+            menu->addSeparator();
+            menu->addAction("Open player in browser", [twitchUsername, this] {
+                this->openChannelIn.invoke(
+                    twitchUsername, FromTwitchLinkOpenChannelIn::BrowserPlayer);
+            });
+            menu->addAction("Open in streamlink", [twitchUsername, this] {
+                this->openChannelIn.invoke(
+                    twitchUsername, FromTwitchLinkOpenChannelIn::Streamlink);
             });
         }
     }